+++ .vscode/launch.json
@@ -109,7 +109,8 @@
         "templedao",
         "traderjoe",
         "truefi",
-        "uniswap",
+        "uniswap-v2",
+        "uniswap-v3",
         "uwu-lend",
         "valas",
         "vector",
