+++ src/adapters/uniswap-v2/index.ts
@@ -1,0 +1,10 @@
+import { Adapter } from '@lib/adapter'
+
+import * as ethereum from './ethereum'
+
+const adapter: Adapter = {
+  id: 'uniswap-v2',
+  ethereum,
+}
+
+export default adapter
